refactor: move all solar-system utils to cel viewer
