+++ apps/web/src/app/asset-loading/load-planets.ts
@@ -1,5 +1,6 @@
 import * as PIXI from 'pixi.js';
-import { spriteConfigs } from '../canvases/solar-system/graphics/config';
+import { spriteConfigs } from '../canvases/celestial-viewer/utils/static-sprite-configs';
+
 import { planetResourcesVar } from '../_state/reactive-variables';
 import { assetLoader } from './asset-loader';
