+++ apps/web/src/app/canvases/celestial-viewer/utils/graphics-utils.ts
