+++ apps/web/src/app/canvases/celestial-viewer/utils/static-sprite-configs.ts
